\name{modelo1}
\alias{modelo1}
\docType{package}
\title{
Modelo 1
}
\description{
Alguns exemplos de parâmetros para explorar o modelo 1.
}
\examples{
Exploração 1: o que representa o r?
Para ganhar entendimento de como o r afeta a equação, podemos alterá-lo enquanto
mantemos os outros parâmetros constantes. Mantenha tmax=10, intervalo=0.20 e 
N₀=10. Altere r conforme os itens seguintes.


1.1 O que acontece se mudar o sinal?
Coloque r=0.5 e clique em apply. Agora faça o mesmo para r=-0.5. E aí?


1.2 O que acontece se o sinal for o mesmo mas o valor diferente?
Coloque r=0.15 e clique em apply. Agora faça o mesmo para r=0.5. E aí?


1.3 E se r for igual a zero?
Coloque r=0 e clique em apply.


Estas explorações devem ser suficientes para capturar as principais característi-
cas do modelo, porém sinta-se à vontade para explorar o que mais quiser!
}
