\name{modelo2}
\alias{modelo2}
\docType{package}
\title{
Modelo 2
}
\description{
Alguns exemplos de parâmetros para explorar o modelo 2.
}
\examples{
Exploração 1: o que representa o r?
Para ganhar entendimento de como o r afeta a equação, podemos alterá-lo enquanto
mantemos os outros parâmetros constantes. Mantenha tmax=20, N₀=10, K=100 e Deixe 
a opção "Gráfico derivada vs N" desmarcada. Altere r conforme os itens seguintes.


1.1 O que acontece se mudar o sinal?
Coloque r=0.5 e clique em apply. Agora faça o mesmo para r=-0.5. E aí?


1.2 O que acontece se o sinal for o mesmo mas o valor diferente?
Para este item coloque tmax=40. Coloque r=0.15 e clique em apply. Agora faça o
mesmo para r=0.5. E aí?


1.3 E se r for igual a zero?
Coloque r=0 e clique em apply.



Exploração 2: o que representa o K?
De maneira análoga, mantenhamos os parâmetros constantes enquanto alteramos o
valor de K. Nesta exploração, o valor de N₀ também será importante, por isso o
alteraremos em conjunto com o K. Mantenha tmax=20 e r=0.5.


2.1 O que acontece se mudarmos o valor de K?
Coloque N₀=10 e k=100 e aperte apply. Agora coloque k=150 e aperte apply. E aí?
Dica: preste atenção nos eixos!

2.2 O que acontece se N₀ começar acima de K?
Agora mantenha k=100 e coloque N₀=50. Agora coloque N₀=150. Tente também N₀=100.



Exploração 3: derivada versus N
Agora faça a exploração do gráfico da derivada em função de N. Para este gráfico
os valores de tmax e N₀ não serão utilizados. O importante aqui é apenas a cara
do grafico, por isso plotá-lo para r=0.5 e K=100 deve ser suficiente. Marque o 
quadro e clique em apply.



Estas explorações devem ser suficientes para capturar as principais característi-
cas do modelo, porém sinta-se à vontade para explorar o que mais quiser!
}
