\name{RcmdrEcoVirtualPlugin-internal}
\title{Internal RcmdrEcoVirtualPlugin objects}
\alias{estdemDb}
\alias{popExp}
\alias{popLog}
\alias{popstrDb}
\alias{propRain}
\alias{intCol}
\alias{resEff}
\alias{resEffcol}
\alias{spDep}
\alias{compDb}
\alias{metacompDb}
\alias{comcompDbox}
\alias{regnichoDbox}
\alias{sucmatrixDbox}
\alias{ColExtDbox}
\alias{arquipDbox}
\alias{bioGeoIslDbox}
\alias{extGameDbox}
\alias{gr.abund}
\alias{hubDbox}
\alias{hubDbox2}
\alias{hubDbox3}
\alias{randWalkDbox}

\description{Internal RcmdrEcoVirtualPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
