\name{RcmdrPlugin.EcoVirtual-package}
\alias{RcmdrPlugin.EcoVirtual-package}
\alias{RcmdrPlugin.EcoVirtual}
\docType{package}
\title{
Rcmdr EcoVirtual Plug-In
}
\description{
This plugin install menus on Rcmdr for use EcoVirtual
}
\details{
\tabular{ll}{
Package: \tab RcmdrPluginEcoVirtual\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-09\cr
License: \tab  GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alexandre Adalardo <aleadalardo@gmail.com>
Maintainer: Alexandre Adalardo <aleadalardo@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}
